add permissions

+++ src/main/java/com/moshna/traker/service/UserService.java
@@ -24,7 +24,7 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-public class UserService  implements UserDetailsService {
+public class UserService implements UserDetailsService {
 
     private final UserRepo userRepo;
     private final ExpenseRepo expenseRepo;
@@ -38,6 +38,10 @@ public class UserService  implements UserDetailsService {
     public static final String USER_ID = "userId";
     public static final String EXPENSES = "expenses";
 
+    public static final String ERROR_MESSAGE = "errorMessage";
+    public static final String PERMISSION_MESSAGE = "You are haven't permissions";
+
+
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         return userRepo.findByUsername(username);
@@ -45,15 +49,22 @@ public class UserService  implements UserDetailsService {
 
     public String getUserList(Model model) {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        if (auth != null && auth.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals(/*Role.ADMIN.name()*/"USER")) || //TODO: использовать Role.USER.name()
-                auth != null && auth.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("MANAGER"))) {
+        if (auth != null && auth.getAuthorities().stream()
+                .anyMatch(a -> a.getAuthority().equals(Role.ADMIN.name())) ||
+                auth != null && auth.getAuthorities().stream()
+                        .anyMatch(a -> a.getAuthority().equals(Role.MANAGER.name()))) {
             List<UserDto> userDtoList = getUserDtoList(getUserList());
+
+            Set<String> roles = Arrays.stream(Role.values())
+                    .map(Role::name)
+                    .collect(Collectors.toSet());
+
             model.addAttribute("users", userDtoList);
             model.addAttribute(USER_ID, getCurrentlyUser().getId());
+            model.addAttribute("allRoles", roles);
             return "userList";
         } else {
-            String message = "You are haven't permissions";
-            model.addAttribute("errorMessage", message);
+            model.addAttribute(ERROR_MESSAGE, PERMISSION_MESSAGE);
             return ERROR;
         }
 
@@ -79,10 +90,10 @@ public class UserService  implements UserDetailsService {
         Set<Role> roleSelected = user.getRoles();
 
         List<String> roleSelectedList = new ArrayList<>();
-        for (Role role: roleSelected) {
+        for (Role role : roleSelected) {
             roleSelectedList.add(role.name());
         }
-        Set<String> roles= Arrays.stream(Role.values())
+        Set<String> roles = Arrays.stream(Role.values())
                 .map(Role::name)
                 .collect(Collectors.toSet());
 
@@ -97,29 +108,38 @@ public class UserService  implements UserDetailsService {
 
     public String userUpdate(Long id, String userName, Map<String, String> form, Model model)
             throws Exception {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth != null && auth.getAuthorities().stream()
+                .anyMatch(a -> a.getAuthority().equals(Role.ADMIN.name())) ||
+                auth != null && auth.getAuthorities().stream()
+                        .anyMatch(a -> a.getAuthority().equals(Role.MANAGER.name()))) {
 
-        User updatedUser = userRepo.findById(id).orElseThrow(()->
-                new Exception("User not found - " + id));
+            User updatedUser = userRepo.findById(id).orElseThrow(() ->
+                    new Exception("User not found - " + id));
 
-       Set<String> roles= Arrays.stream(Role.values())
-                .map(Role::name)
-                .collect(Collectors.toSet());
+            Set<String> roles = Arrays.stream(Role.values())
+                    .map(Role::name)
+                    .collect(Collectors.toSet());
 
-        updatedUser.getRoles().clear();
+            updatedUser.getRoles().clear();
 
-        for(Map.Entry<String, String> entry: form.entrySet()) {
-            String key = entry.getKey();
-            if(roles.contains(key)) {
-                updatedUser.getRoles().add(Role.valueOf(form.get(key)));
+            for (Map.Entry<String, String> entry : form.entrySet()) {
+                String key = entry.getKey();
+                if (roles.contains(key)) {
+                    updatedUser.getRoles().add(Role.valueOf(form.get(key)));
+                }
             }
+
+            model.addAttribute(USER_ID, getCurrentlyUser().getId());
+            model.addAttribute(EXPENSES, expenseService.getExpenseDtoList(expenseRepo.findAllByUserId(id)));
+
+            updatedUser.setUsername(userName);
+            userRepo.save(updatedUser);
+            return USER_EXPENSE;
+        } else {
+            model.addAttribute(ERROR_MESSAGE, PERMISSION_MESSAGE);
+            return ERROR;
         }
-
-        model.addAttribute(USER_ID, getCurrentlyUser().getId());
-        model.addAttribute(EXPENSES, expenseService.getExpenseDtoList(expenseRepo.findAllByUserId(id)));
-
-        updatedUser.setUsername(userName);
-        userRepo.save(updatedUser);
-        return USER_EXPENSE;
     }
 
     public String addExpense(
@@ -129,21 +149,40 @@ public class UserService  implements UserDetailsService {
             String comment,
             Model model
     ) {
-        expenseService.addExpense(description,comment, price, id);
-        return getUserExpenses(model);
+        User user = getCurrentlyUser();
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth != null && auth.getAuthorities().stream()
+                    .anyMatch(a -> a.getAuthority().equals(Role.ADMIN.name())) ||
+                     id == user.getId()) {
+            expenseService.addExpense(description, comment, price, id);
+            return getUserExpenses(model);
+        } else {
+            model.addAttribute(ERROR_MESSAGE, PERMISSION_MESSAGE);
+            return ERROR;
+        }
     }
 
     public String userExpenseDetails(Long id, Model model) throws Exception {
+        User user = getCurrentlyUser();
+        Expense expense = expenseRepo.findById(id).orElseThrow(() -> new Exception("Expense not found - " + id));
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth != null && auth.getAuthorities().stream()
+                .anyMatch(a -> a.getAuthority().equals(Role.ADMIN.name())) ||
+                expense.getUserId() == user.getId()) {
 
-        Expense expense = expenseRepo.findById(id).orElseThrow(()->new Exception("Expense not found - " + id));
-        model.addAttribute(USER_ID, getCurrentlyUser().getId());
-        model.addAttribute(EXPENSES, expense);
-        return EXPENSE_DETAILS;
+
+            model.addAttribute(USER_ID, user.getId());
+            model.addAttribute("expense", expense);
+            return EXPENSE_DETAILS;
+        } else {
+            model.addAttribute(ERROR_MESSAGE, PERMISSION_MESSAGE);
+            return ERROR;
+        }
     }
 
     public String expenseUpdate(Long id, String description, String comment,
                                 BigDecimal price, Model model) throws Exception {
-        Expense expense = expenseRepo.findById(id).orElseThrow(()->new Exception("Expense not found - " + id));
+        Expense expense = expenseRepo.findById(id).orElseThrow(() -> new Exception("Expense not found - " + id));
 
         List<ExpenseDto> expenseDtoList =
                 expenseService.getExpenseDtoList(expenseRepo.findAllByUserId(getCurrentlyUser().getId()));
@@ -160,8 +199,9 @@ public class UserService  implements UserDetailsService {
     }
 
     public String removeUser(Long id, Model model) {
+        //TODO: you can't remove yourself
         List<Expense> expenses = expenseRepo.findAllByUserId(id);
-        for (Expense e: expenses) {
+        for (Expense e : expenses) {
             expenseRepo.deleteById(e.getId());
         }
         userRepo.deleteById(id);
@@ -173,12 +213,21 @@ public class UserService  implements UserDetailsService {
     }
 
     public String removeExpense(Long id, Model model) {
-        expenseRepo.deleteById(id);
+        User user = getCurrentlyUser();
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth != null && auth.getAuthorities().stream()
+                .anyMatch(a -> a.getAuthority().equals(Role.ADMIN.name())) ||
+                id == user.getId()) {
+            expenseRepo.deleteById(id);
 
-        model.addAttribute(USER_ID, id);
-        model.addAttribute(EXPENSES, expenseService.getExpenseDtoList(
-                expenseService.getExpenseList(getCurrentlyUser().getId())));
-        return USER_EXPENSE;
+            model.addAttribute(USER_ID, id);
+            model.addAttribute(EXPENSES, expenseService.getExpenseDtoList(
+                    expenseService.getExpenseList(getCurrentlyUser().getId())));
+            return USER_EXPENSE;
+        } else {
+            model.addAttribute(ERROR_MESSAGE, PERMISSION_MESSAGE);
+            return ERROR;
+        }
     }
 
     public String filterByDates(Long id, String fromDate, String toDate, Model model) {
@@ -188,7 +237,7 @@ public class UserService  implements UserDetailsService {
         model.addAttribute("averageExpense", expenseService.getAverageExpense(
                 expenseFilteredDtoList));
         model.addAttribute(EXPENSES, expenseFilteredDtoList);
-        model.addAttribute(USER_ID,id);
+        model.addAttribute(USER_ID, id);
         return USER_EXPENSE;
     }
 
@@ -199,19 +248,19 @@ public class UserService  implements UserDetailsService {
             currentUserName = authentication.getName();
         }
         User currentUser = new User();
-        List<User> userList =  userRepo.findAll();
-        for (User user: userList) {
-            if(user.getUsername().equals(currentUserName)) {
+        List<User> userList = userRepo.findAll();
+        for (User user : userList) {
+            if (user.getUsername().equals(currentUserName)) {
                 currentUser = user;
                 break;
             }
         }
-        return  currentUser;
+        return currentUser;
     }
 
     public List<UserDto> getUserDtoList(List<User> userList) {
         List<UserDto> userDtoList = new ArrayList<>();
-        for (User user: userList) {
+        for (User user : userList) {
             userDtoList.add(UserMapping.toUserDto(user));
         }
         return userDtoList;
